third_party/WebKit/LayoutTests/fast/canvas/canvas-empty-image-pattern.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<canvas id="canvas" width="100" height="100"></canvas>

<script>
test(function(t) {
    var ctx = document.getElementById("canvas").getContext("2d");
    var pattern = ctx.createPattern(new Image, "repeat")
    assert_equals(pattern, null);
}, 'Test if filling a pattern with an empty image does not crash and returns a null pattern.');
</script>
